requirements.txt
# core
pymupdf>=1.24.0          # PDF text + metadata extraction
pdfplumber>=0.7.6        # tables
python-dotenv>=0.21.0    # secure .env loading
google-generativeai>=0.5.0   # Gemini API SDK
langchain-google-genai>=0.1.0 # optional, LangChain connector for Gemini
chromadb>=0.4.0          # vector DB (local start)
pandas>=1.5.0            # for tables
tiktoken>=0.4.0          # (tokenizer; useful even if not strictly needed for Gemini)
streamlit>=1.25.0        # simple UI demo
arxiv>=1.4.8             # optional (bonus)
pytest>=7.0.0            # tests
